Transport: ensure gateway_id is always a string
It must be a string because it is used as a client id for
mqtt connection.
If set from cmd line, it is already ensured but not if it comes from
yaml config file or if id is automatically generated with getNode().

Correct fix will be to parse all yaml settings with argparse too.
It will be done later.