Single Dockerfile for multi-arch builds
This commit removes the need of keeping two Dockerfiles for
different architectures.

However, it still relies on docker-compose to build release
versions or the .ci/build-images.sh for continuous integration
builds.

In terms of release builds, these will only be built when a tag
is given to ci/travis. This is in preparation for the multi-arch
push to docker hub.

There is still a need in the future to accomodate other OSes inside
the container folder, such as Alpine.