Build optimizations for Docker (#60)
This commit removes the need of keeping two Dockerfiles for
different architectures. 

Similarly as before, the Dockerfiles are kept in two folders, a stable 
for builds based on https://github.com/wirepas/manifest and a dev folder
for builds based on the current content of the repo (+ master of c-mesh-api). 
A commons folder contains templates and the docker-entrypoint script.

Release builds are now only triggered when a tag is given to CI system, which
is meant to prepare the future push from CI to dockerhub.

There is still a need in the future to accommodate other OSes inside
the container folder, such as Alpine.

* Update readme (#63) Closes #62
* Update wirepas_messaging version Bump to rc3. See #38