Handle unset boolean parameters
Currently any key associated with a boolean value is added to the
argument list to be fed into argparse.

However, this will cause the parameter to be set to True even when
its value on the file is False.

To overcome that, when reading from the yaml file, any key that has
a value of False will be skipped. Thus we rely that the default
value handles it correctly and that store_true is the only
boolean related action allowed.

If a store_false is neeed in the future one could espace it with a
name rule, such as --no-feature-name.