Docker builds: manifest handling
This commit fixes the path issues in the docker-compose for the
stable versions and updates the location of where to
source the python version from (__about__.py instead of __init__.py)

Similar issue was still visible in .ci/build-images which is now
fixed as well.

The manifest branch is now default to an actual branch and not
a tag. The reason for this is that we are planning to use
branches on wirepas/manifest to control the gateway version.