Run yaml parameters through argparse (#69)
The type of input parameters is not respected when reading them
from a yaml file. This happens due to the fact that the dictionary
obtained from yaml is used to overwrite entries in the argument
list.

This change calls the parse_args from the argparse to avoid
issues with the value's type. 

Command line parameter will always take precedence over file 
settings.

Closes #67