Docker-builds: metadata fix and template updates (#84)
This commit fixes the path issues in the docker-compose for the
stable versions and updates the location of where to
source the python version from (__about__.py instead of __init__.py)

Similar issue was still visible in .ci/build-images which is now
fixed as well.

The manifest branch is now default to an actual branch and not
a tag. The reason for this is that we are planning to use
branches on wirepas/manifest to control the gateway version.

The entrypoint template was using arguments that are to be deprecated
in version 2.0. This changes updates the template to have a 1:1
match with the 1.2.0 parameters.

Closes #65