Docker deploy updates (#108)
The fecth artifact was still pulling images from the edge tag
instead of the tagged build. With this change the script
relies on TRAVIS_TAG to set the build tag.

The manifest creation was also failing as the images need to
be pushed to the repository prior to its creation.