Avoid MQTT reconnect() loop
This endless loop is happening when the MQTT broker closes the connection early,
due to invalid credentials.
This patch aborts the MQTT processing if a disconnect is detected
before a successful connection.

Note that it may be a bug in the broker, but this behavior is compatible
with the MQTT spec:
"The Server MAY check that the contents of the CONNECT Packet meet any
further restrictions and MAY perform authentication and authorization
checks. If any of these checks fail, it SHOULD send an appropriate
CONNACK response with a non-zero return code as described in section 3.2
and it MUST close the Network Connection."