Sink_manager: Handle additionnal Exception
When reading param from sinkService, the sinkService is supposed
to generate a DbusError if not able to read it. It generates
a GLib.Error in python.
Even if I was not able to reproduce, a customer reported that
this Dbus exception can be promoted as an AttributeError instead.

So catching it to avoid raising this exception at a higer level.