SinkService: generate StackStarted message on manual restart
StackStarted was only generated when the node was booting
as it was mapped to the boot indication from dual mcu api.

But this indication doesn't happen during a manual start.
And this signal can be important for any transport logic.
As the stack could be started by a different procees
(different transport) other transport were not notified.

So generate the signal on an explicit start.