Python interface tests
This changes allows pytest to execute the transport service's
tests and cleanups the service tests to allow for a custom
command.

The boolean arguments are problematic at the moment, since it
is only possible to set them to True (for False, you must
rely on the default value).

A separate PR will be raised to improve how we handle this.