mqtt_wrapper: fix buffering issues during reconnect
first issue: while reconnecting, the wrapper thread never
read the internal Selectable Queue. The queue size is not
limited BUT the socket used for signaling has a memory limit,
and will start to be blocking after (approximately) 400
published messages.

The patch change the way the signal socket is used.


second issue: The messages stored in the internal Selectable Queue
are not used for the publish queue size computation. It prevents the
black hole detection to work while reconnection is in progress.

The patch update the publish queue size computation.