Fix a blackhole detection regression (WiP)
Due to this regression, it is possible to see a cost sink increase for
a short duration (1 second), and with no reason.


A previous modification of the 'publish_queue_size' computation, that fix
an issue with buffer threshold detection when the network connection is
lost, could break the publish delay detection in some conditions.

Note that this regression is racy, and not so easy to catch.
So for the record, this is what the log looks like when it happens:

2020-03-05 09:14:49,289 | [INFO] wirepas_gateway@transport_service.py:714:Request to send data
2020-03-05 09:14:49,296 | [INFO] wirepas_gateway@transport_service.py:349:Increasing sink cost of all sinks
2020-03-05 09:14:50,327 | [INFO] wirepas_gateway@transport_service.py:362:Connection is back, decreasing sink cost of all sinks

1) "Request to send data" is not important, it just means that a message is published (SendDataResponse)
2) Then the blackhole detection run while the message is still in the wrapper queue.
The publish_queue_size is not empty but publish timestamp is not updated,
so the old timestamp is used to compute the delay.
3) The message in the wrapper queue is processed, the check works again.



This patch updates the 'last_published_packet_s' property to take into account the
modification of 'publish_queue_size'