try to reduce MQTT messages reordering after a reconnection
When the MQTT wrapper reconnect to the broker, the in-flight messages
(stored in the mqtt client queue) are processed after the new messages
coming from the wrapper queue.

This leads to a harmless reordering of messages, that some customers
want to avoid anyway.


This patch changes to the order of the MQTT messages processing, to keep
the orignal message order in this use-case.