SinkService: automatic uart speed
If uart baudrate is not given, test the list of bitrates.
For now it is 125000, 115200 and 1000000.
The list can be extended if at some point we have different speed
in use.
With the timeouts, it takes arround 1s per speed to check a baudrate.

c-mesh-api submodule sha1 is updated at same time as it is required
for this feature to work.