fix a start/stop race in dbus_client
The self.loop.quit() behavior is not consistent with the documentation.
A following call to self.loop.run() should return but it does not.

Due to this issue, if the mqtt_wrapper get a fast connection failure
at the startup, the TransportService will stay alive without any
connection to a MQTT broker.


This patch uses the MainLoop to call quit() and avoid the race between
the calls to run() and quit()