Implementation of the RTC service on top of DBUS
This service sends periodically time to the network.
It allows Wirepas gateway to send a global clock to the devices in the network with a good precision and accuracy without having specific hardware.
The service gets the global time in UTC timezone from an ntp server and spreads the value to the network.