Transport: first connection not mandatory
Before this change, in case of a failing first connection to broker,
the transport service was stopped (and restarted automatically by
systemd or docker daemon).
It was resulting of messages sent by sink-service(s) being lost as no one
was listenning it.
So now, transport will use its internal mechanism to retry the connection
and buffers the received packet until the connection is successful.
Same behavior as when connection is lost.

It is particularly important at boot time when internet connectivity
is not availble immediatelly.