Status update: fix unsent online status
To avoid too much publish on status topic (that may trigger more traffic from client),
there is a cache an the new status is not sent if same as last one (it is retained on broker anyway)

But the cache was not handling the disconnect properly. In case of disconnect, the "last will" update the
status but the gateway do not see it and still think that the retained status is the one from its cache.

So the fix is just to clean the cache in the onConnect callback