Enable clearing of sink scratchpad
According to the gateway to backend API, scratchpad should be cleared if
the scratchpad bytes are not present in the upload scratchpad request.

This change enables clearing of the scratchpad in order to make it
easier to test OTAP functionality of the gateway.

Additionally, in order to prevent accidental clearing of the scratchpad,
the clearing is only done if the scratchpad bytes field is missing (i.e.
it should fail if an empty file is uploaded) and the sequence id is set
to UINT32_MAX.