LG-71 Respond to SetScratchpadTargetAndActionReq with invalid parameters
Unlike most of the other messages, parameter validation for
SetScratchpadTargetAndActionReq is done on the wirepas-mesh-messaging
library side and not in transport service.

Before this change, from_payload() method would raise an exception in
those cases, but a response could not be sent because the req_id was not
known. Newer version of the wirepas-mesh-messaging library raises an
InvalidMessageContents exception which includes the request header, so
it is possible to build a response message.