LG-90 Missing error code mappings for CDD operations
Getting list of configuration data items would fail with the
APP_RES_OPERATION_NOT_SUPPORTED error code only if an older,
incompatible version of stack is used with dualmcu_app.

Since getting the list of items is done regularly for the status message
in the gateway, such return code is not really an error.