LG-93 Prevent status thread from crashing
If a sink is removed when the status thread is reading the sink
configuration, unhandled dbus exceptions caused the status thread to
crash.

This fix is done in multiple layers. First, missing exception handling
is added to get_scratchpad_status to prevent it from crashing if a sink
is not existing when it is executed. Secondly, "stored_scratchpad" and
"processed_scratchpad" entries are not returned if they are empty (empty
entries caused problems with message serialization). These changes alone
prevent the thread from crashing. Finally, in the status thread,
_set_status() is called around a try block to catch any other exception
that might have been missed.